+++ src/Kiota.Builder/KiotaBuilder.cs
@@ -875,7 +875,7 @@ public partial class KiotaBuilder
         {
             Access = AccessModifier.Private,
             Name = "urlTemplate",
-            DefaultValue = $"\"{currentNode.GetUrlTemplate()}\"",
+            DefaultValue = $"\"{currentNode.GetUrlTemplateForPathItem()}\"",
             ReadOnly = true,
             Documentation = new()
             {
@@ -1473,7 +1473,10 @@ public partial class KiotaBuilder
             var operationUrlTemplate = currentNode.GetUrlTemplate(operationType);
             if (!operationUrlTemplate.Equals(parentClass.Properties.FirstOrDefault(static x => x.Kind is CodePropertyKind.UrlTemplate)?.DefaultValue?.Trim('"'), StringComparison.Ordinal)
                 && currentNode.HasRequiredQueryParametersAcrossOperations())// no need to generate extra strings/templates as optional parameters will have no effect on resolved url.
+            {
                 generatorMethod.UrlTemplateOverride = operationUrlTemplate;
+                executorMethod.UrlTemplateOverride = operationUrlTemplate;
+            }
 
             var mediaTypes = (schema, operation.Responses is null) switch
             {